+++ src/tooling/piral-cli/src/common/config.ts
@@ -71,7 +71,7 @@ export const config: PiralCliConfig = rc('piral', {
   url: undefined,
   cert: undefined,
   npmClient: 'npm',
-  bundler: 'webpack',
+  bundler: 'webpack5',
   piletApi: '/$pilet-api',
   validators: {},
   schemaVersion: 'v2',