+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.15.0 (tbd)
 
 -  Moved templates to dedicated repository (#458)
+-  Changed the default bundler to be `piral-cli-webpack5` (#469)
 
 ## 0.14.9 (January 21, 2022)